into to double
Next. Look for input validation of it's not a number and allow a number like 2.5 input, change c and f(credits) to a short or float or double. Might have to getline and if it's a number break else loop again. Look for those alternative solution to the wrong inputs on accident and implement them for every input validation.